refactor: [GNS-255] UI Layer for Search
