+++ src/components/view/search/not-found/index.ts
@@ -1,0 +1,1 @@
+export * from "./NotFound";
