+++ src/layouts/error/error-404/Error404Layout.styles.ts
@@ -1,0 +1,11 @@
+import { innerLayoutCss } from "@/styles/css/inner-layout";
+import styled from "styled-components";
+
+export const Container = styled.main`
+  width: 100%;
+  flex: 1;
+`;
+
+export const InnerLayout = styled.div`
+  ${innerLayoutCss}
+`;
